hard error when no mempools are configured
