+++ crates/driver/example.toml
@@ -11,6 +11,9 @@ endpoint = "http://localhost:1235"
 relative-slippage = "0.1"
 account = "0x0000000000000000000000000000000000000000000000000000000000000002"
 
+[[submission.mempool]]
+mempool = "public"
+
 [contracts] # Optionally override the contract addresses, necessary on less popular blockchains
 gp-v2-settlement = "0x9008D19f58AAbD9eD0D60971565AA8510560ab41"
 weth = "0xC02aaA39b223FE8D0A0e5C4F27eAD9083C756Cc2"
