+++ crates/driver/src/domain/mod.rs
@@ -1,6 +1,7 @@
 pub mod competition;
 pub mod eth;
 pub mod liquidity;
+mod mempools;
 pub mod quote;
 
-pub use {competition::Competition, liquidity::Liquidity};
+pub use {competition::Competition, liquidity::Liquidity, mempools::Mempools};
